+++ src/linear_prime_sieve.rs
@@ -1,11 +1,11 @@
 /// compute least prime factor table and prime numbers list.
-pub fn linear_prime_sieve(size: usize) -> (Vec<Option<u64>>, Vec<u64>) {
+pub fn linear_prime_sieve(size: usize) -> (Vec<Option<u32>>, Vec<u32>) {
     let mut lpf = vec![None; size];
     let mut prime_numbers = Vec::with_capacity(size);
     for i in 2..size {
         if lpf[i].is_none() {
-            lpf[i] = Some(i as u64);
-            prime_numbers.push(i as u64);
+            lpf[i] = Some(i as u32);
+            prime_numbers.push(i as u32);
         }
         for &p in &prime_numbers {
             if p > lpf[i].unwrap() || p as usize * i >= size {
@@ -29,7 +29,7 @@ mod tests {
         };
         const K: usize = 1 << 10;
         let lpf_ans = least_prime_factor_table(K);
-        let primes_ans = find_prime_numbers(K as u64);
+        let primes_ans = find_prime_numbers(K as u32);
         assert_eq!(
             (lpf_ans, primes_ans),
             linear_prime_sieve(K)
